# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anatomiya",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/term": {
            "get": {
                "tags": [
                    "Terms"
                ],
                "summary": "All Terms",
                "description": "Hamma Terminlarlarni ko'rish",
                "operationId": "term_index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Term"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Terms"
                ],
                "summary": "new Trems add",
                "description": "Yangi Terms qoshish",
                "operationId": "term_store",
                "requestBody": {
                    "description": "lesson save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "keyword_latin",
                                    "keyword_ru",
                                    "keyword_uz",
                                    "keyword_en",
                                    "description_uz",
                                    "description_ru",
                                    "description_en"
                                ],
                                "properties": {
                                    "keyword_latin": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Harley"
                                    },
                                    "keyword_ru": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Augustus"
                                    },
                                    "keyword_uz": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Augustus"
                                    },
                                    "keyword_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Augustus"
                                    },
                                    "description_uz": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Description"
                                    },
                                    "description_ru": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Description"
                                    },
                                    "description_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Term"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/term/{id}": {
            "get": {
                "tags": [
                    "Terms"
                ],
                "summary": "Show term",
                "description": "bitta termda hamma malumotlarini ko'rsatadi",
                "operationId": "term_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Term"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Terms"
                ],
                "summary": "Update branch",
                "operationId": "term_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "keyword_latin",
                                    "keyword_uz",
                                    "keyword_ru",
                                    "keyword_en",
                                    "description_uz"
                                ],
                                "properties": {
                                    "keyword_latin": {
                                        "type": "string",
                                        "example": "keyword_latin"
                                    },
                                    "keyword_ru": {
                                        "type": "string",
                                        "example": "keyword_ru"
                                    },
                                    "keyword_uz": {
                                        "type": "string",
                                        "example": "keyword_uz"
                                    },
                                    "keyword_en": {
                                        "type": "string",
                                        "example": "keyword_en"
                                    },
                                    "description_uz": {
                                        "type": "text",
                                        "example": "description"
                                    },
                                    "description_ru": {
                                        "type": "text",
                                        "example": "description"
                                    },
                                    "description_en": {
                                        "type": "text",
                                        "example": "description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Term"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Terms"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "term_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Term"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "title": "ResponseError",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "text",
                            "example": "Not Found"
                        }
                    }
                },
                "type": "object"
            },
...
            "Term": {
                "title": "Term",
                "required": [
                    "name_uz",
                    "name_ru",
                    "image",
                    "parent_id"
                ],
                "properties": {
                    "name_uz": {
                        "type": "string"
                    },
                    "name_ru": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "parent_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "student": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "bearer_token": {
            "type": "apiKey",
            "description": "Enter token in format (Bearer <token>)",
            "name": "Authorization",
            "in": "header"
        }
    }
}